+++ packages/no-modal/src/base/plugin/IPlugin.ts
@@ -1,8 +1,9 @@
 import { type SafeEventEmitter, WhiteLabelData } from "@web3auth/auth";
 
 import { type Analytics } from "../analytics";
-import { CHAIN_NAMESPACES } from "../chain/IChainInterface";
-import { type IWeb3AuthCore } from "../core/IWeb3Auth";
+import { CHAIN_NAMESPACES, type CustomChainConfig } from "../chain/IChainInterface";
+import type { CONNECTOR_STATUS_TYPE } from "../connector/constants";
+import type { Connection, IConnector } from "../connector/interfaces";
 import { type WALLET_CONNECTOR_TYPE } from "../wallet";
 
 export const PLUGIN_NAMESPACES = {
@@ -39,12 +40,20 @@ export const WALLET_PLUGINS = {
   ...SOLANA_PLUGINS,
 } as const;
 
+export interface IWeb3AuthPluginHost {
+  currentChain: CustomChainConfig | undefined;
+  connectedConnectorName: WALLET_CONNECTOR_TYPE | null;
+  status: CONNECTOR_STATUS_TYPE;
+  connection: Connection | null;
+  getConnector(connectorName: WALLET_CONNECTOR_TYPE): IConnector<unknown> | null;
+}
+
 export interface IPlugin extends SafeEventEmitter {
   name: string;
   status: PLUGIN_STATUS_TYPE;
   SUPPORTED_CONNECTORS: WALLET_CONNECTOR_TYPE[];
   pluginNamespace: PluginNamespace;
-  initWithWeb3Auth(web3auth: IWeb3AuthCore, whiteLabel?: WhiteLabelData, analytics?: Analytics): Promise<void>;
+  initWithWeb3Auth(web3auth: IWeb3AuthPluginHost, whiteLabel?: WhiteLabelData, analytics?: Analytics): Promise<void>;
   connect(): Promise<void>;
   disconnect(): Promise<void>;
   cleanup(): Promise<void>;
